requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
requests==2.32.3
google-genai==0.3.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1
pandas==2.2.2
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.5